Ensuring BindingsClash message sent to app when bindings clash
